libaktualizr: package_manager: validate server uri
This early validation works as a buffer before the underlying ostree functions
are hit. If we let this validation go to the lower levels it becomes quite
hard to link against different ostree/glib variations.

For example, using the glib/ostree combination available on bookworm provides
the following error from glib if a bad uri is passed through

	URI is not absolute, and no base URI was provided

Whereas using bullseye we get the previous error from ostree instead

	Failed to parse uri: bad-url

Signed-off-by: Leonardo Held <leonardo.held@toradex.com>